rename package to not include underscores
