--- src/main/kotlin/dev/domnikl/schema_registry_gitops/SchemaParseException.kt
@@ -1,3 +1,0 @@
-package dev.domnikl.schema_registry_gitops
-
-class SchemaParseException(message: String) : RuntimeException(message)
