--- src/main/kotlin/dev/domnikl/schema_registry_gitops/ServerVersionMismatchException.kt
@@ -1,3 +1,0 @@
-package dev.domnikl.schema_registry_gitops
-
-class ServerVersionMismatchException(message: String, cause: Throwable) : RuntimeException(message, cause)
