--- src/main/kotlin/dev/domnikl/schema_registry_gitops/state/Dumper.kt
@@ -1,20 +1,0 @@
-package dev.domnikl.schema_registry_gitops.state
-
-import dev.domnikl.schema_registry_gitops.SchemaRegistryClient
-import dev.domnikl.schema_registry_gitops.State
-import dev.domnikl.schema_registry_gitops.Subject
-
-class Dumper(private val client: SchemaRegistryClient) {
-    fun dump() = State(
-        client.globalCompatibility(),
-        client.subjects().map { subject ->
-            val schema = client.getLatestSchema(subject)
-            Subject(
-                subject,
-                client.compatibility(subject),
-                schema,
-                schema.references()
-            )
-        }
-    )
-}
