+++ src/main/kotlin/dev/domnikl/schemaregistrygitops/Compatibility.kt
@@ -1,0 +1,11 @@
+package dev.domnikl.schemaregistrygitops
+
+enum class Compatibility {
+    NONE,
+    BACKWARD,
+    FORWARD,
+    FULL,
+    BACKWARD_TRANSITIVE,
+    FORWARD_TRANSITIVE,
+    FULL_TRANSITIVE
+}
