+++ src/main/kotlin/dev/domnikl/schemaregistrygitops/SchemaParseException.kt
@@ -1,0 +1,3 @@
+package dev.domnikl.schemaregistrygitops
+
+class SchemaParseException(message: String) : RuntimeException(message)
