+++ src/main/kotlin/dev/domnikl/schemaregistrygitops/ServerVersionMismatchException.kt
@@ -1,0 +1,3 @@
+package dev.domnikl.schemaregistrygitops
+
+class ServerVersionMismatchException(message: String, cause: Throwable) : RuntimeException(message, cause)
